backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
google-cloud-storage==2.14.0
boto3==1.29.7
httpx==0.25.2
python-dotenv==1.0.0
python-multipart==0.0.6

